Split the test code version of raw_ffi out
This commit starts splitting up x11rb::xcb_ffi::raw_ffi. The first step
is to move the implementation used for tests to a separate file.
